Few more places using size_t instead of int

+++ test/test-json.c
@@ -66,7 +66,7 @@ int main() {
 
   str = xjsonEscape(specials, 0);
   if(strlen(str) != 2 * strlen(specials)) {
-    fprintf(stderr, "ERROR: '%s' has length %d, expected %d\n", str, (int) strlen(str), (int) strlen(specials));
+    fprintf(stderr, "ERROR: '%s' has length %zu, expected %zu\n", str, strlen(str), strlen(specials));
     return 1;
   }
